+++ src/model/IceProtectionRocket.java
@@ -1,7 +1,7 @@
 package model;
 
 public class IceProtectionRocket extends Rocket {
-    public IceProtectionRocket(double range,double height,Airspace a){
+    public IceProtectionRocket(double range,int height,Airspace a){
         super( range, height,"IPR",a);
     }
 }
