+++ src/model/MilitaryRocket.java
@@ -3,7 +3,7 @@ package model;
 public class MilitaryRocket extends Rocket {
     private boolean inland;
 
-    public MilitaryRocket(double range, double height, boolean i, Airspace a) {
+    public MilitaryRocket(double range, int height, boolean i, Airspace a) {
 
         super(range, height, "MiR", a);
         inland = i;
