+++ src/model/PessengerAirplane.java
@@ -9,7 +9,7 @@ public class PessengerAirplane extends Aircraft {
     private double maximumWeightOfLuggage;
 
 
-    public PessengerAirplane(String model, double height, Map characteristics, List persons, int numOfSeats, double max, Airspace a) {
+    public PessengerAirplane(String model, int height, Map characteristics, List persons, int numOfSeats, double max, Airspace a) {
         super(model, height, characteristics, persons, "PrA", a);
         numberOfSeats = numOfSeats;
         maximumWeightOfLuggage = max;
