+++ src/model/PessengerHelicopter.java
@@ -8,7 +8,7 @@ public class PessengerHelicopter extends Aircraft {
     private int numberOfSeats;
 
 
-    public PessengerHelicopter(String model, double height, Map characteristics, List persons, int numOfSeats, Airspace a) {
+    public PessengerHelicopter(String model, int height, Map characteristics, List persons, int numOfSeats, Airspace a) {
         super(model, height, characteristics, persons, "PrH", a);
         numberOfSeats = numOfSeats;
     }
