+++ src/model/TransportHelicopter.java
@@ -7,7 +7,7 @@ public class TransportHelicopter  extends Aircraft{
     private String cargo;
     private double maximumWeightOfCargo;
 
-    public TransportHelicopter(String model, double height, Map characteristics, List persons, String cargo,double max,Airspace a){
+    public TransportHelicopter(String model, int height, Map characteristics, List persons, String cargo,double max,Airspace a){
         super(model,height,characteristics,persons,"TrH",a);
         this.cargo=cargo;
         maximumWeightOfCargo=max;
